Add configuration option for controlling file IO flushing behavior
This is just a slightly simpler alternative to PR #187 with a more generic
name to leave room for future developments. The thing is, flushing out IO
is something rpm should really do always on *some* level, but on rotational
disks doing so on this simple per-file level is so horribly expensive that
it's simply not practical. SSD's are an entirely different story however,
and on some workloads regularly flushing out IO has other benefits (which
is what PR #187 is actually about) even if you dont actually care about the
durability aspects.

Documentation is intentionally left scarce as this is not at a stage
where we'd want to actually advertise it.

(cherry picked from commit 4087530f0fcbb14167be8296957e44e6ffc97579)