And another test-suite related distcheck regression fix...
Commit ec0d4d80d8ba1c1541018915ea47aca7aa956251 added a bunch of
perfectly innocent looking rpm queries to the rpmbuild test suite,
which however end up breaking "make distcheck". Really.

The problem is that these are executed without the testsuite run()
wrapper function so the rpm running environment isn't quite right
and because of that, when run deep from distcheck in a crazy
prefix config (or something like that), an rpmdb instance gets left
over where it shouldn't be, and that makes distcheck unhappy.
And that in turn tends to make your head hurt.