Fix st_size on non-archive use of rpmfiStat()/rpmfilesStat()
In archives, hardlinked files are zero-sized except for the last file
of the hardlinked set which carries the actual content. For other
purposes we want the actual size, hardlink or not. So in rpmfilesStat()
always just report the size from the header and fixup the size to
when iterating over archives from rpmfiStat() instead.