Don't initialize return values on rpmReadHeader() entry
For consistency with rpmReadPackageFile() (which changed behavior
similarly in the grand package read unification) and common practise,
only set return values pointers when we have something to return.
Other than that, everybody gets to initialize their own variables.

Also don't dereference msg, we really dont need to know whether it
points to a NULL string ornot in order to set it.