Fix "substatus" (eg NOKEY) regression on package read (#330)
Commit 3ccd9ccf50c8d643f9dcd956155720018429beed introduced a regression
in the cases where package is successfully read (ie a header is returned)
but "there is a but", such as a missing pubkey. We need to return the
substatus instead of just "yes yes" in that case.

Keep track of the substatus in through the callback and return it
in case everything else was okay to restore former behavior.