Avoid local path in testsuite expected results, oops
Commit 5bd9f5f580723098c14f6a5c6cb398813967f738 was relying on a local
path on my laptop, obviously failing anywhere else. In most tests
those messages are not an issue but since python tests run outside
the test-root, they expose local paths. For regular cases we could
just tell autotest to [ignore] the result, but this gets lost somewhere
in the ~three layers of quotes and whatnot of passing python foo around,
in order to be able to escape handle [] sanely in python code. Lovely.
So shut redirect rpmlog() to /dev/null instead.