Add support for new virtual file attribute "%artifact"
This can be used to differentiate files that are not natural parts of
packages but created as by-products of our processing so they're easy
to filter out of queries.

Possible candidates include build-ids, byte compiled files etc, but this
nothing is automatically marked as artifact in this commit.

(cherry picked from commit 6f1e75ddd2c67eb8b43608c03bf0cc895612e6fe)