Mark build-id and dwz entries in packages as %artifact files
In worst-case scenarios (packages with lots of binaries but little
else), the build-id entries can dominate the query output to the
point its hard to see the actual content. Marking these things as
%artifact entries (artificial by-products of our processing) allows
them to be easily filtered out from queries, eg:

  $ rpm -qpl can-utils-20170830git-1.fc27.x86_64.rpm | wc -l
  76
  $ rpm -qpl --noartifact /can-utils-20170830git-1.fc27.x86_64.rpm | wc -l
  28

Besides build-ids in the main package(s), this also marks the build-id
and dwz-related entries in the debuginfo packages as %artifacts. The
debuginfo files exist solely for carrying debuginfo data so its perhaps
arguable whether these should be %artifacts in there too or only in the main
packages, but ... if the very same thing is artifact in one then is it
not also in the other? Additionally, all these build-ids and the dwz
entries are in hidden directories on the filesystem even in the debuginfo
packages which suggests they might be best swiped out of sight.

(cherry picked from commit 80934099292c9e5acfb98216bfcd6ef73bf7df13)