Implement include-filtering for verification too
Since it comes at very nearly zero cost now, support include-filtering
(--configfiles etc) for verify too.

The actual functionality is little more than a single if-continue
in the verify code, what unfortunately makes this commit so "big" is
rearranging the popt tables to have separate file selection options
table that is common for both query and verify, plus update the
manual to reflect that, plus testcases for the functionality.