Fix error message for rpmlock (#355), by removing dead code
rpmlock_acquire() failed to set errno, when it failed to acquire a
write-lock on a lock which was opened as read-only.  errno was printed in
the error message, so the error message was arbitrary.  Typically the
previous errno happened to be the right one, but if STDIN was not a TTY,
then it was not.  (Due to an intervening call to isatty()).

A one-line fix to set errno appropriately would not fix #355.  Because the
appropriate errno for performing write operations on a read-only handle
is EBADF.  This is too generic to tell the user what the real problem is
(i.e. that rpm was run without the necessary privilege to write to the lock
file).

This obstacle would disappear if we could avoid the automatic fallback
from a read-write open() to a read-only open().  It turns out that
rpmlock.h never exported a read-only lock feature, so this was never used.

Remove the code for automatic fallback to read-only and read-only locks.
This makes #355 go away.