Move %transfiletriggerun after problem checking and fingerprinting
There's no reason for this to be as evil as %pretrans, it's used for
entirely different purposes such as unregistering a whole bunch of
modules at once on the uninstall-transaction, and we don't want that
to happen if there were file conflicts present.

Additionally, file triggers should not fire on skipped / non-installed
files. This is currently not taken into account and not changed here,
but fingerprinting results are a pre-requisite for doing that.

(cherry picked from commit d83edaa2188ecd13fabbf1d779763dcc17bf1b93)