Expect failure for invalid directory symlink fails when run as root
The test-suite doesn't generally expect or like to be run as root, but
that's how our CI runs it at the moment. And with the new symlink
verification rules from commit c815822c8bdb138066ff58c624ae83e3a12ebfa9,
when the "invalid" link is unexpectedly created by root this breaks our
testcase (but not the actual behavior of the code).