Fix inode handling for zero-sized files
Both the prevSize and fixupSize arguments in rpmtsUpdateDSI use zero
as "unset" value. But that means that the inode count is not changed
for zero-sized files. So we imcrement the values in the arguments and
later decrement them again in rpmtsUpdateDSI.

(cherry picked from commit 791cf86205348270eedcca5912cbbc2f68de0de0)