Improve hardlink handling in the disk space calculation
Before this commit, rpm simply did not take hardlinks into account
when calculating disk space requirements. This made it fail
spectaculary for packages that contain a high number of hardlinks,
like glibc-locale.

We now "bind" the file size to the last hardlink member. This
is still not 100% correct as the last file may not get installed
in some cases or only the last file may be replaced. But it is
still much better than what rpm did before.