Eliminate file trigger second argument (aka $2)
File triggers have gotten both $1 and $2 as zero always, which doesn't
make a whole lot of sense. For $1 we know what to do (it should be the
same as with all other scriptlet types) but $2 is not so obvious. For
regular triggers it's the number of remaining instances of the target
package, but with file triggers there are any number of target packages
and it's not clear what the most useful thing would be. Better not
pass it at all to make it free for future use once we figure what to
do with it.