rpmrc: Add architecture compatibility mapping between aarch64 and arm64
AArch64 has another common name for the architecture: arm64. And while compilers,
the kernel, and RPM use "aarch64", a number of language ecosystems use "arm64".

In addition, Debian packages use the "arm64" architecture name to refer to "aarch64",
and AArch64 packages that are translated from DEB to RPM format tend to not be
installable because there's not a sufficient architecture mapping.

This change adds the "arm64" identifier as an alias for "aarch64" so that this works
as expected, rather than breaking on unknown architectures.