Allow verification callbacks to signal stopping, document the callback
Move the rc code from return value to a parameter, change to a pointer
to allow modification from the callback. Make the message modifiable
from callback too if only for consistency. Change the parameter names
on all callback users to force all relevant places to be updated.
With all this, a verification callback now force verification to abort
by returning 0. No actual behavior is changed in this commit though.