Make rpmErase() exit code resemble rpmInstall() exit code
rpmtsRun() only returns the number of failures if problems prevented
the transaction from starting at all. All other errors such as failure
to remove due to scriptlet exit are simply returned as -1. When that
gets returned to shell it ends up being 255 which makes even less sense,
so lets at least mimic the way rpmInstall() returns the number of
*attempted* operations for the failure. It's not *right* or *sensible*
of course, but then nothing about our exit codes is.