Allow packagers to specify individual scriptlets as critical
Since rpm 4.6.0 (commit 9cfb380cc4529ea6b9314cc67819069c8e9c1a23) we've
filtered out most scriptlet failures as non-critical: most of the time
these scriptlets failing are just somebody forgetting to add ||: to
the last line or so. And even in case of real errors, the damage
is already done by the time we get to post scriptlets - the package
is installed or removed and no exit code can undo that. Distro people
are generally quite happy with this, but there's been a constant demand
to bring the former behavior back from other types of users.

We dont want to bring back the mess of duplicate packages left behind on
upgrades due to silly missing ||: or fluff like info register/unregister
failure because somebody ran with --nodocs etc. But clearly some scriptlets
are more important that others, for example kernel install/update failing
to create an initrd (RhBug:1394770) could be considered as such. So lets
give the choice to packagers - who else is supposed to know if their
scriptlets are worth failing install/erase for?

Since the infrastructure for this was added in commit
d8e46a1a48e4ae3d06fb10fcb2615a06be169dd6 already, this is little more
than a oneliner to add '-c' option to spec scriptlet sections, eg
'%post -c' to mark that %post scriptlet critical. Additionally
the error codes from post-transaction scriptlets were not passed
through to rpmtsRun() exit code before this, because there simply
was no need to do so. And of course a testcase  duplicating the scriptlet
failure tests with scriptlets marked critical.