Further relax type checking among different string tag types (#414)
In some cases i18n strings are just strings in packages (at least
summary in rhel-6 created src.rpms), no reason to fail because of that,
just assume all string tag types are compatible. i18n-strings are not
truly equal to the other types but this is close enuf for the purpose.