Fix a macro end detection if the first char of a macro line is '}' (#401)
The algorithm that detects the end of a macro counts for all macro chars their level of nested. It iteratively moves from the first char of the macro further.
Before the patch the rpm function does not implement this algorithm correctly. After finishing with the last char of a macro line, it additionally skips one char of the next line. The first char in the line was omitted.

Probably affects every rpm version from this millenium.

(cherry picked from commit 2456c128948ac4ade1dbd4a61198a278c23e40be)