Streamline digest/signature info initialization
rpmsinfoInit() never gets called with an unknown value as it's us
calling it based on values in our built-in tables (it used to be
different though). In fact it mostly gets called while we're walking
the very same table we're then looking up again in rpmsinfoInit()...
makes no sense whatsoever. Change the internal API to take the
table values directly, and handle the lookup in the only case that
needs it (payload digest).

Not supposed to change any functionality except make things a little
saner and faster too.