Separate signature/digest verification from the digest finalization
Adjust the API a bit to make things clearer/saner (hopefully): we now init
and finalize *ranges*, and verification is an independent, non-destructive
operation that can be performed on given *types* of items, allowing
separate digest and signature verification and can be performed multiple
times without rereading data (for eg added keys). At least in theory...
which is to say no such things are done here, this is not supposed to
change any visible functionality.

For now at least, sort the items in rpmvsVerifyItems() by range to
preserve traditional rpm -Kv output.