Preliminaries for reporting non-present verification items
Remember RPMRC_NOTFOUND results in verify set, but ignore them in
actual results for now. The check for verifiable types never really
belonged to rpmsinfoDisabled() as it's a different kind of check,
and here we see why - we NEVER want to add non-verifiable elements
to verify set, but for the next steps we'll need to handle disabled
items as if they were not present - ie NOTFOUND.

No functionality is supposed to change here.