Add popt-based options --setcaps and --restore (RhBug:1550745)
--setcaps resets capabilities of package files.

Because the opotion is popt-based it has some drawbacks -
it does not know about %ghost files, file states, etc.

--restore regenerate owner, group, permissions and capabilities
of package files.

It uses 3 already defined options in the correct order.
rpm --setugids is used first of all, because it may change permissions
or capabilities. rpm --setperms is used before rpm --setcaps, but the
opposite order is right too.