Implement configurable, mandatory signature/digest verify level for rpm -K
Traditionally rpm has verified whatever digests and signatures are present
(unless disabled) but there hasn't been any way to *require* either type
of verification, and making it all too easy to fool rpm into accepting
a tampered package whose signatures and digests have just been stripped
off and content modified. This commit adds support for required
verification level, possible levels being "none" (for traditional behavior
and default for now), "digest", "signature" and "all" (for both signatures
and digests).

What it means is that the specified level of validation is required for
the entire package file: eg on digest level, both the header and the
payload must have been validated with a digest. All digest algorithms are
considered equal for this purpose to abstract out version differences:
eg all valid rpm packages have *some* digest covering the entire package.
Traditionally the only digest covering the payload is the md5sum, but
packages built with rpm >= 4.14.0 have a separate payload digest which
can be used for the purpose and md5sum disabled (which would also deny
install of older packages).

Similarly (and more importantly) signatures can now be also required.
As a special case, the new payload digest added in 4.14.x is in the
main header which, if signed, makes the payload digest signature strength
too. Which would make it possible to drop V3 header+payload signatures
from package signing (which would make signing much faster) and still
have signature strength coverage for the entire package.

It's worth noting that when a verification level is active, --nodigest
and --nosignature seem to behave sort of counter-intuitively: these disablers
tell rpm not to try verification of those items, but the verification
level specifies what *must* be verified for a pass, so eg if signature
verify level is active, rpm -K --nosignatures will *always fail* because no
signatures could be verified.

This only implements the "policy" for the rpm -K style signature checking
path, installation is not covered yet as that's quite a different beast:
we don't want to enforce this at rpmReadPackage() level because of
API issues and assumptions (eg you couldn't read a package to discover
what key it was signed with, in order to import the key required to
successfully verify said package)