Really fix CVE-2017-7500. Ugh.
This is what you get for rushing a fix and relying on a test-suite that
cannot test ownership changes :( The idea behind passing the stat struct
from the package in commit f2d3be2a8741234faaa96f5fd05fdfdc75779a79
perhaps was that we'd actually want the owner of the target directory in
the package, not disk. But it's not the target directory we're passing
there (we can't easily get that data from here) so the test is bogus
and thus doesn't fix the actual problem. Just use the target directory
permissions on disk since that's the best we have, and remove the now
unused argument from fsmVerify() again.

Thanks to Michael Schroeder and SUSE QA team for noticing and reporting!

(cherry picked from commit d3fae43615f4e925e21fd12a96974a85542f13b9)