Fix dubious code in rpmscript.c which could lead to uninitalized variable use
Commit 9c3e5de3240554c8ea1b29d52eeadee4840fefac,  "Factor out
and unify setting CLOEXEC" moved code which set the xx variable.

This accidentally caused problems in Yocto Project use of rpm
since we don't use the setenv call in the function and this meant
scripts would fail to run with exit code 127. Since that code is
within its own code block its hard to spot and the code is confusing
at best.

Looking back into history, the code which originally set xx upon
chdir was removed and as I understand it, xx is a convention used to
ignore return values for linting purposes. If error checking is
desired in this function, there are much clearer ways it could be done.

It would seem to make sense to simply remove the dubious check
entirely.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>