Unbreak numberless %patch from the autonumering changes
Commit 53036261dda93885106901a983976eb0e9ee6ea2 incorrectly claims
that numberless Patch: and Source: referred to number zero internally,
instead they were special cased to INT_MAX which allowed differentiating
from zero and the cost of ugly hacks here and there, and making referring
to the patch/source by %{PATCH/SOURCE<N>) macro rather impossible.

Undo the rest of the INT_MAX magic too and make numberless %patch really
equal patch number zero. This makes it work for the common case, for
example in Fedora's > 19000 specs there is not a single case where
the former special INT_MAX magic would've been necessary. If this change
breaks somebody's spec, I'll promise to light a candle for it on a
dark and cold winter night, we have plenty of those here in Finland.