Adjust testsuite for CVE-2017-7500 verify tweak, deep sigh
Commit 10bd0373179f203633eb8e4e392b94926d306b7a broke the testsuite
as we can't test anything wrt file ownership on disk. The test only
"worked" because it looked at the packaged value, but that's not
really right either. How hard can it be? Sigh...

Anyway, we can only test for this now if we're root, and make the
test actually work for *that* case.