Add rpmcliVSFlags which is populated with the actual vsflags from popt
This is a backport of commit 80f8d6376066a3e2e303965d7f0e462d0f2bb8bb
but since that commit is an API/ABI break, we need to do this in
addition to the existing mechanism instead of just ditching the old one
and support both. That is, in the unlikely event that somebody actually
uses the rpmcli-side of things, but you never know.

This is only needed to minimize differences to master to make package
verify backport easier.