macros: make unversioned python macros not expand to empty when the executable is not installed
In Fedora the removal of /usr/bin/python in favour of explicit
/usr/bin/python2 and /usr/bin/python3 is happening now. Unless
python-unversioned-command.rpm is installed, /usr/bin/python does not
exists. This means that packages use
%python_{sitelib,sitearch,version}, they get a warning in the build
log, and the macro evaluates to an empty string.
It seems better to keep the unexpanded macro instead, as if it wasn't
defined at all.

See https://bugzilla.redhat.com/show_bug.cgi?id=1585626#c4.