macros: Use full %___build_template.
DRY! Instead of 'aborting' the multipart '%___build_template' for stage
scripts after the 'pre' step and fiddling in the 'body' and 'post' parts
in 'doScript()' separately, we make it complete with 'body' and 'post'.
The details will be filled dynamically in one fell swoop.

The interesting part is the invocation of '%___build_template' when the
stage scripts are created: Instead of writing the content of the 'sb'
buffer to the script file directly, we ship this content to the
'%___build_body' macro, which is referenced in the middle of the
'%___build_template' and hence will be expanded and shipped to the script
file automagically.

In order to avoid regressions, the content of 'sb', which gets expanded
when the specfile is first digested, has to be 'macro-escaped' before
the snippet can be processed by 'rpmExpand()' again, i.e., any existing
'%' character has to be doubled.