Clean up and improve rpmlog error handling + reporting
Centralize the log failure printing to a helper function, handle
the common conditions there. Remember the last error and only
print a new one if the error differs to avoid unnecessary flooding.
Finally, make the actual message more concise and mark it for i18n.
Oh and add a testcase as well.

This will still try to perror() in vain even if it's stderr that's
returning the errors in the first place, but maybe that's perror()'s
problem and not ours.

(cherry picked from commit 41cf6f9c8038a99e1e4930a8e70289ca5c8eff5e)