fileattrs: Don't scan libraries in glibc auto-search dirs for Provides
Those libraries are meant to be runtime replacements of the ones in the
regular $libdir, based on CPU features at runtime. If a packager decides
to split them into a separate sub-package, the dependency resolver
shouldn't be deciding to install that instead of the base package.

But we do need to process them for Requires.