debugedit: support compressed debug sections
If -gz=zlib is used at compile time the debug sections will be zlib
compressed and need to be decompressed before interpretting. Without
this compressed data will blindly be used as if it were uncompressed
giving inconsistent results.

Support for handling compressed sections was added to elfutils in
version 0.165 so we only do this if we are on that version or newer.